Add shared setupTests file, move workflowstage mock to separate mock file. Start work to convert snapshot tests to enzyme.
