+++ RetrospectiveExtension.Frontend/components/__mocks__/WorkflowStage.tsx
@@ -1,0 +1,6 @@
+export const mockWorkflowStage = {
+    display: 'test',
+    value: 'Collect',
+    isActive: true,
+    clickEventCallback: () => {}
+};
