+++ RetrospectiveExtension.Frontend/components/__tests__/feedbackItem.test.tsx
@@ -7,7 +7,7 @@ import { mockEnv } from '../__mocks__/config/environment';
 import { mockCore } from '../__mocks__/azure-devops-extension-api/Core/Core';
 import { mockCommon } from '../__mocks__/azure-devops-extension-api/Common/Common';
 import { MockSDK } from '../__mocks__/azure-devops-extension-sdk/sdk';
-import {WorkflowPhase} from '../../interfaces/workItem';
+import { WorkflowPhase } from '../../interfaces/workItem';
 import { v4 as uuid } from 'uuid';
 import FeedbackItem from '../feedbackItem';
 import FeedbackColumn from '../feedbackColumn';
@@ -20,31 +20,6 @@ Enzyme.configure({ adapter: new Adapter() });
 // Base render constants, these may change if the FeedbackItem component is changed.
 const childDialogCount = 5;
 const voteButtonCount = 2;
-
-// Mock Environment
-jest.mock('../../config/environment', () => { return mockEnv; });
-
-// Mock Azure DevOps Extension SDK
-jest.mock('azure-devops-extension-sdk', () => { return MockSDK; });
-
-// Mock Azure DevOps Extension API
-jest.mock('azure-devops-extension-api/Core', () => { return mockCore; });
-
-jest.mock('azure-devops-extension-api/Core/CoreClient', () => {});
-jest.mock('azure-devops-extension-api/WebApi', () => {});
-jest.mock('azure-devops-extension-api/WorkItemTracking', () => {});
-jest.mock('azure-devops-extension-api/WorkItemTracking/WorkItemTracking', () => {});
-jest.mock('azure-devops-extension-api/WorkItemTracking/WorkItemTrackingClient', () => {
-  const mockWorkItemTrackingClient = {
-    WorkItemTrackingRestClient: {},
-  };
-
-  return mockWorkItemTrackingClient;
-});
-
-jest.mock('azure-devops-extension-api/Common', () => {
-  return mockCommon;
-});
 
 const testTeamId = uuid();
 const testBoardId = uuid();
