+++ RetrospectiveExtension.Frontend/package.json
@@ -30,6 +30,7 @@
     "babel-jest": "^27.4.6",
     "classnames": "2.3.1",
     "copy-to-clipboard": "3.3.1",
+    "enzyme-to-json": "^3.6.2",
     "file-saver": "2.0.5",
     "jest-junit-reporter": "^1.1.0",
     "moment": "2.29.1",
@@ -92,6 +93,9 @@
     },
     "transformIgnorePatterns": [
       "node_modules/(?!(azure-devops-ui|azure-devops-extension-sdk|azure-devops-extension-api)/)"
+    ],
+    "setupFiles": [
+      "./components/__tests__/setupTests.tsx"
     ],
     "moduleNameMapper": {
       "office-ui-fabric-react/lib/(.*)$": "office-ui-fabric-react/lib-commonjs/$1"
